backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
imbalanced-learn==0.11.0
